+++ source/theme/rules/app/core/components/buttonControl.ts
@@ -13,6 +13,8 @@ export const getRules = (themeContext: ThemeContext): object => {
     "button.separator": styles.fill.onAccent.pri,
     "checkbox.background": styles.fill.control.rest,
     "checkbox.border": styles.stroke.control.default,
+    "checkbox.disabled.background": "#FF0000FF", // 🔵 Unknown attribute
+    "checkbox.disabled.foreground": "#FF0000FF", // 🔵 Unknown attribute
     "checkbox.foreground": styles.fill.text.active,
     "checkbox.selectBackground": styles.fill.control.rest,
     "checkbox.selectBorder": styles.stroke.control.default,