+++ source/theme/rules/code/syntax/core/scopes/10.ts
@@ -49,5 +49,23 @@ export const getRules = (themeContext: ThemeContext): object => {
         fontStyle: "",
       },
     },
+    // Braces
+    {
+      scope: [
+        "meta.brace.round",
+        "punctuation.definition.arguments",
+        "punctuation.definition.begin.bracket.round",
+        "punctuation.definition.end.bracket.round",
+        "punctuation.definition.group.begin",
+        "punctuation.definition.group.end",
+        "punctuation.definition.parameters",
+        "punctuation.section.function.begin.bracket.round",
+        "punctuation.section.function.end.bracket.round",
+      ],
+      settings: {
+        foreground: color,
+        fontStyle: "",
+      },
+    },
   ];
 };